+++ R/deprivation_decile.R
@@ -16,7 +16,9 @@
 #' @seealso \code{\link{indicators}} for indicator lookups,
 #'   \code{\link{profiles}} for profile lookups
 #'   \code{\link{indicator_metadata}} for the metadata for each indicator and
-#'   \code{\link{area_types}} for area types and their parent mappings
+#'   \code{\link{area_types}} for area types and their parent mappings and
+#'   \code{\link{category_types}} for category lookups and
+#'   \code{\link{indicator_areatypes}} for indicators by area types lookups
 
 deprivation_decile <- function(AreaTypeID = 102, Year = 2015) {
         if (!(Year %in% c(2010,2015))) {