+++ R/retrieve_data.R
@@ -1,33 +1,46 @@
 #' @importFrom jsonlite fromJSON
 #' @importFrom utils read.csv
-
-retrieve_indicator <- function(IndicatorIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
+#' @importFrom httr GET content set_config config
+#' @importFrom readr cols
+retrieve_indicator <- function(IndicatorIDs, ProfileIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
         path <- "https://fingertips.phe.org.uk/api/"
+        set_config(config(ssl_verifypeer = 0L))
         fingertips_data <- data.frame()
-        # total <- length(IndicatorIDs) * length(ChildAreaTypeIDs) * length(ParentAreaTypeIDs)
-        # i <- 0
-        for (IndicatorID in IndicatorIDs) {
+        for (i in 1:length(IndicatorIDs)) {
+                IndicatorID <- IndicatorIDs[i]
                 for (ChildAreaTypeID in ChildAreaTypeIDs) {
                         for (ParentAreaTypeID  in ParentAreaTypeIDs) {
-                                # i <- i + 1
-                                # pb <- txtProgressBar(min = 0,
-                                #                      max = total,
-                                #                      style = 3)
-                                dataurl <- paste0(path,
-                                                  sprintf("all_data/csv/by_indicator_id?indicator_ids=%s&child_area_type_id=%s&parent_area_type_id=%s",
-                                                          IndicatorID,ChildAreaTypeID,ParentAreaTypeID),
-                                                  "&include_sortable_time_periods=yes")
-                                fingertips_data <- rbind(read.csv(dataurl),
+                                if (missing(ProfileIDs)){
+                                        dataurl <- paste0(path,
+                                                          sprintf("all_data/csv/by_indicator_id?indicator_ids=%s&child_area_type_id=%s&parent_area_type_id=%s",
+                                                                  IndicatorID, ChildAreaTypeID, ParentAreaTypeID),
+                                                          "&include_sortable_time_periods=yes")
+                                } else {
+                                        ProfileID <- ProfileIDs[i]
+                                        if (is.na(ProfileID)) {
+                                                dataurl <- paste0(path,
+                                                                  sprintf("all_data/csv/by_indicator_id?indicator_ids=%s&child_area_type_id=%s&parent_area_type_id=%s",
+                                                                          IndicatorID,ChildAreaTypeID,ParentAreaTypeID),
+                                                                  "&include_sortable_time_periods=yes")
+                                        } else {
+                                                dataurl <- paste0(path,
+                                                                  sprintf("all_data/csv/by_indicator_id?indicator_ids=%s&child_area_type_id=%s&parent_area_type_id=%s&profile_id=%s",
+                                                                          IndicatorID, ChildAreaTypeID, ParentAreaTypeID, ProfileID),
+                                                                  "&include_sortable_time_periods=yes")
+                                        }
+                                }
+                                fingertips_data <- rbind(dataurl %>% GET %>% content("parsed", type = "text/csv", encoding = "UTF-8", col_types = cols()),
                                                          fingertips_data)
-                                # Sys.sleep(0.1)
-                                # setTxtProgressBar(pb, i)
                         }
                 }
         }
-        #close(pb)
         return(fingertips_data)
 }
 
+#' @importFrom jsonlite fromJSON
+#' @importFrom utils read.csv
+#' @importFrom httr GET content set_config config
+#' @importFrom readr cols
 retrieve_domain <- function(DomainIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
         path <- "https://fingertips.phe.org.uk/api/"
         fingertips_data <- data.frame()
@@ -38,7 +51,7 @@ retrieve_domain <- function(DomainIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
                                                   sprintf("all_data/csv/by_group_id?child_area_type_id=%s&parent_area_type_id=%s&group_id=%s",
                                                           ChildAreaTypeID,ParentAreaTypeID,DomainID),
                                                   "&include_sortable_time_periods=yes")
-                                fingertips_data <- rbind(read.csv(dataurl),
+                                fingertips_data <- rbind(dataurl %>% GET %>% content("parsed", type = "text/csv", encoding = "UTF-8", col_types = cols()),
                                                          fingertips_data)
                         }
                 }
@@ -46,6 +59,10 @@ retrieve_domain <- function(DomainIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
         return(fingertips_data)
 }
 
+#' @importFrom jsonlite fromJSON
+#' @importFrom utils read.csv
+#' @importFrom httr GET content set_config config
+#' @importFrom readr cols
 retrieve_profile <- function(ProfileIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
         path <- "https://fingertips.phe.org.uk/api/"
         fingertips_data <- data.frame()
@@ -56,7 +73,7 @@ retrieve_profile <- function(ProfileIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
                                                   sprintf("all_data/csv/by_profile_id?child_area_type_id=%s&parent_area_type_id=%s&profile_id=%s",
                                                           ChildAreaTypeID,ParentAreaTypeID,ProfileID),
                                                   "&include_sortable_time_periods=yes")
-                                fingertips_data <- rbind(read.csv(dataurl),
+                                fingertips_data <- rbind(dataurl %>% GET %>% content("parsed", type = "text/csv", encoding = "UTF-8", col_types = cols()),
                                                          fingertips_data)
                         }
                 }