+++ tests/testthat/test-deprivation.R
@@ -15,9 +15,9 @@ test_that("the dimensions of deprivation decile function are as expected", {
         expect_equal(dim(dep_101_2010), c(326, 3))
 })
 
-exp_classes <- c("factor","numeric","numeric")
+exp_classes <- c("character","numeric","numeric")
 names(exp_classes) <- c("AreaCode", "IMDscore", "decile")
-test_that("the class of columns returned are factor-numeric-numeric", {
+test_that("the class of columns returned are character-numeric-numeric", {
         expect_equal(sapply(dep_default,class), exp_classes)
         expect_equal(sapply(dep_2010,class), exp_classes)
         expect_equal(sapply(dep_101,class), exp_classes)