+++ vignettes/fingertips_paper_captions.R
@@ -2,4 +2,4 @@
 
 fig_caps <- captioner(prefix = "Figure")
 
-fig_caps(name = "life_exp_birth", "Life expectancy at birth by Index of Multiple Deprivation score, Unitary Authority and Upper Tier Local Authority, England 2012-14")
+fig_caps(name = "life_exp_birth", "Life expectancy at birth by Index of Multiple Deprivation score, Upper Tier Local Authority, England 2012-14")