+++ NEWS.md
@@ -1,3 +1,21 @@
+# fingertipsR 0.1.1
+
+* `select_indicators()` allows user to point and click to select indicators
+
+* stringsAsFactors parameter available in `fingertips_data()`
+
+* automatically filter for `CategoryType = FALSE` in `fingertips_data()` - this can be set to `TRUE` if needed
+
+* rank of area and polarity of indicator returned from `fingertips_data()` where `rank = TRUE` (polarity can also be found in `indicator_metadata()`)
+
+* `fingertips_redred` highlights which areas are statistically different to comparator *and* trending in the wrong direction
+
+* `category_types()` lookup function to support ordering where categories exist (eg, deprivation decile)
+
+* `areatypes_by_indicators()` to help users determine which indicators are available for each area type (and vice versa)
+
+* A new vignette demonstrating how some of the new functions can be used
+
 # fingertipsR version 0.1.0
 
 This package allows the user to retrieve tables of: